:hammer: Move into different mailer classes. Add staging and production actionmailer configs
