+++ app/views/goodcity_order_mailer/send_appointment_confirmation_email.html.erb
