+++ app/views/goodcity_order_mailer/send_order_submission_delivery_email.en.html.erb
