+++ app/views/goodcity_order_mailer/send_order_submission_pickup_email.zh-tw.html.erb
